+++ packages/anoncreds/src/formats/index.ts
@@ -2,6 +2,7 @@ export * from './AnonCredsDidCommCredentialFormat'
 export { AnonCredsDidCommCredentialFormatService } from './AnonCredsDidCommCredentialFormatService'
 export * from './AnonCredsDidCommProofFormat'
 export { AnonCredsDidCommProofFormatService } from './AnonCredsDidCommProofFormatService'
+export { AnonCredsLinkSecretBindingProvider } from './AnonCredsLinkSecretBindingProvider'
 export { DataIntegrityDidCommCredentialFormatService } from './DataIntegrityDidCommCredentialFormatService'
 export * from './LegacyIndyDidCommCredentialFormat'
 export { LegacyIndyDidCommCredentialFormatService } from './LegacyIndyDidCommCredentialFormatService'
