+++ packages/didcomm/src/modules/credentials/formats/dataIntegrity/index.ts
@@ -1,2 +1,4 @@
 export * from './DidCommDataIntegrityCredentialFormat'
+export * from './DidCommDataIntegrityCredentialFormatService'
+export * from './DidCommDataIntegrityLinkSecretBinding'
 export * from './dataIntegrityExchange'
